Move repositories&AppDbContext to OnlineShop.Data

+++ OnlineShop.WebApi/Controllers/ProductController.cs
@@ -1,21 +1,21 @@
 using Microsoft.AspNetCore.Mvc;
+using OnlineShop.Data.Repositories;
 using OnlineShop.Models;
-using OnlineShop.WebApi.Data;
 
 namespace OnlineShop.WebApi.Controllers;
 
 [Route("products")]
 public class ProductController : ControllerBase
 {
-    private readonly IRepository<Product> _productRepository;
+    private readonly IProductRepository _productRepository;
 
-    public ProductController(IRepository<Product> productRepository)
+    public ProductController(IProductRepository productRepository)
     {
-        _productRepository = productRepository;
+        _productRepository = productRepository ?? throw new ArgumentNullException(nameof(productRepository));
     }
 
     [HttpGet("get_all")]
-    public async Task<IReadOnlyList<Product>> GetProducts(CancellationToken cancellationToken)
+    public async Task<IReadOnlyList<Product>> GetProducts(CancellationToken cancellationToken = default)
     {
         var products = await _productRepository.GetAll(cancellationToken);
         return products;
@@ -29,20 +29,23 @@ public class ProductController : ControllerBase
     }
 
     [HttpPost("add")]
-    public async Task AddProduct(Product product, CancellationToken cancellationToken)
+    public async Task AddProduct(Product product, CancellationToken cancellationToken = default)
     {
+        if (product == null) throw new ArgumentNullException(nameof(product));
         await _productRepository.Add(product, cancellationToken);
     }
 
     [HttpPut("update")]
-    public async Task Edit(Product product, CancellationToken cancellationToken)
+    public async Task Edit(Product product, CancellationToken cancellationToken = default)
     {
+        if (product == null) throw new ArgumentNullException(nameof(product));
         await _productRepository.Update(product, cancellationToken);
     }
 
     [HttpPost("delete")]
-    public async Task DeleteProduct(Product product, CancellationToken cancellationToken)
+    public async Task DeleteProduct(Product product, CancellationToken cancellationToken = default)
     {
+        if (product == null) throw new ArgumentNullException(nameof(product));
         await _productRepository.Delete(product, cancellationToken);
     }
 }
